fast int16

+++ matrix.h
@@ -26,16 +26,16 @@ void imm_matrix_dump(struct imm_matrix *, FILE *);
 TEMPLATE float imm_matrix_get_score(struct imm_matrix const *x,
                                     struct imm_cell const y)
 {
-  unsigned const row = y.row % IMM_MATRIX_NROWS;
-  unsigned const col = (unsigned)(x->state_col[y.state] + y.len);
+  uint_fast8_t const row = y.row % IMM_MATRIX_NROWS;
+  uint_fast16_t const col = (uint_fast16_t)(x->state_col[y.state] + y.len);
   return imm_matrixf_get(&x->score, row, col);
 }
 
 TEMPLATE void imm_matrix_set_score(struct imm_matrix *x,
                                    struct imm_cell const y, float const score)
 {
-  unsigned const row = y.row % IMM_MATRIX_NROWS;
-  unsigned const col = (unsigned)(x->state_col[y.state] + y.len);
+  uint_fast8_t const row = y.row % IMM_MATRIX_NROWS;
+  uint_fast16_t const col = (uint_fast16_t)(x->state_col[y.state] + y.len);
   imm_matrixf_set(&x->score, row, (unsigned)col, score);
 }
